backend/requirements.txt
# 软件建设方案 AI 生成系统 - Python 依赖
# 版本：V2.0

# Web 框架
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# 数据库
SQLAlchemy==2.0.23

# 文档处理
python-docx==1.1.0

# HTTP 请求
requests==2.31.0

# 模板引擎
Jinja2==3.1.2

# 异步任务（可选）
celery==5.3.4
redis==5.0.1

# 测试
pytest==7.4.3
pytest-cov==4.1.0
